Prisma, optimisitc UI and modal/toast restructure
