+++ pages/api/createReview.js
@@ -2,17 +2,23 @@ import prisma from "../../lib/prisma"
 
 export default async function handle(req, res) {
     await prisma.account.update({
+        // Match user ID
         where: { id: 1 },
         data: {
             reviews: {
                 create: {
                     rating: req.body.rating,
                     Cafe: {
-                        create: {
-                            googlePlaceID: req.body.googlePlaceID,
-                            latitude: req.body.latitude,
-                            longitude: req.body.longitude,
-                            name: req.body.name,
+                        connectOrCreate: {
+                            where: {
+                                googlePlaceID: req.body.googlePlaceID,
+                            },
+                            create: {
+                                googlePlaceID: req.body.googlePlaceID,
+                                latitude: req.body.latitude,
+                                longitude: req.body.longitude,
+                                name: req.body.name,
+                            },
                         }
                     }
                 }
